+++ sampledb/models/migrations/user_add_is_hidden.py
@@ -1,11 +1,11 @@
 # coding: utf-8
 """
-Add is_readonly column to users table.
+Add is_hidden column to users table.
 """
 
 import os
 
-MIGRATION_INDEX = 12
+MIGRATION_INDEX = 13
 MIGRATION_NAME, _ = os.path.splitext(os.path.basename(__file__))
 
 
@@ -16,12 +16,12 @@ def run(db):
         FROM information_schema.columns
         WHERE table_name = 'users'
     """).fetchall()
-    if ('is_readonly',) in client_column_names:
+    if ('is_hidden',) in client_column_names:
         return False
 
     # Perform migration
     db.session.execute("""
         ALTER TABLE users
-        ADD is_readonly BOOLEAN NOT NULL DEFAULT(FALSE)
+        ADD is_hidden BOOLEAN NOT NULL DEFAULT(FALSE)
     """)
     return True