+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Version 0.9
 Currently in development.
 
 - Added readonly users
+- Allow hiding users
 
 Version 0.8.1
 -------------